requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# LangGraph and AI
langgraph==0.0.62
# Align with langgraph which requires langchain-core >=0.2,<0.3
langchain==0.2.12
langchain-core==0.2.34
langchain-openai==0.1.7
openai==1.40.6

# Computer Vision
opencv-python==4.8.1.78
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
Pillow==10.1.0

# YouTube capture
yt-dlp==2023.11.16

# Data processing
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2

# Web UI
jinja2==3.1.2
aiofiles==23.2.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Utilities
requests==2.31.0
tqdm==4.66.1
